Implemented regex matching on the URIs for extra control
The URIs passed to the second array in the constructor of SelectivePluginLoader
can now use regex to give extra control on when a plugin is enabled or
disabled. Previously, due to the partial, it was impossible (for example)
to specify that the plugin should only be shown on the home page only.

This change is backwards compatible, so existing installations should
not be affected.